autoscaler: use int64 accumulators in doOptimize direction
Summing int32 replica counts into int32 accumulators can overflow
if totals exceed ~2.1B, producing an incorrect direction and wrong
interval selection. Promote the accumulators to int64 to match the
direction type and avoid overflow.

+++ pkg/autoscaler/controller/autoscale_controller.go
@@ -345,14 +345,14 @@ func (ac *AutoscaleController) doOptimize(ctx context.Context, binding *workload
 	}
 	// Compute direction — compare only targets that exist in both
 	// replicasMap and recommendedInstances, matching the update loop below.
-	var currentSum, recommendedSum int32
+	var currentSum, recommendedSum int64
 	for name, current := range replicasMap {
 		if recommended, ok := recommendedInstances[name]; ok {
-			currentSum += current
-			recommendedSum += recommended
+			currentSum += int64(current)
+			recommendedSum += int64(recommended)
 		}
 	}
-	direction := int64(recommendedSum) - int64(currentSum)
+	direction := recommendedSum - currentSum
 
 	// Do update replicas
 	for _, param := range optimizer.Meta.Config.Params {
